snap: make `snap download` use snapd if available
The snap download command will now try to talk to snapd
directly. There is also an additional "--direct" switch now to
force going to the store directly and it will automatically fallback
if no snapd is available.

This also adds more tests to `snap download` but its still not greatly
tested.

Based on #7622