cmd/snap-confine: stop being setgid root
Historically snapd has been setuid root and setgid root. During the
openSUSE security review it was pointed out that the setgid root aspect
is redundant because snap-confine can call setegid(0) to achieve the
same effect.

This patch performs exactly that change. Packaging is updated in
distributions that model permissions in the packaging meta-data.

This has the consequence of fixing incorrect group ownership of
the following files:
 - *.fstab files in /run/snapd/ns
 - *.lock files in /run/snapd/lock
 - cgroup in /sys/fs/cgroup/devices

I think that dropping/raising effective group is error-prone and
since we don't permanently drop the permission until the very end,
is not an effective security measure.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>