overlord/devicestate: make gadget-assets-update task work during remodel, pick update policy
Update the gadget-assets-update task to work in remodel scenario. Tweak the
approach to finding the state information since the name of the snap we're
updating to may not be the same as the name of the current gadget snap.

Pick the right policy for the remodel or regular boot assets update.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>