interfaces: de-duplicate emitted update-ns profiles
It was brought to our attention that apparmor_parser cannot de-duplicate
the following profile efficiently, unlike all the other rules:

    mount foo,
    mount foo,

Internally apparmor_parser first de-duplicates some things that are
identical.  This step is efficient in complexity and is not a problem.
The next step constructs a state machine based on all the rules that
remain and minimises it.  Unfortunately this step is exponential in
memory so while the resulting binary profile is efficient the means of
computing one is not.

To combat this, this patch switches apparmor.Specification to use
OrderedSet to store update-ns snippets. Existing buffered emitters are
replaced by a de-duplicating emitter. Since the automatically generated
update-ns rules are very repetitive, that last step ensures that
actually generated profile text does not contain the repetitions that
would stress apparmor_parser.

Fixes: https://bugs.launchpad.net/snapd/+bug/1848567
Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>