tests: update mount-ns after addition of /etc/systemd/user
We have recently added a bind mount for /etc/systemd/user to support
user-session services. This was not reflected in the mount-ns test
because core part of those tests are still disabled due to leaking
elements in two core tests.

This patch updates the test to pass if explicitly enabled.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>