tests: verify host is not affected by mount-ns tests
The mount-ns test performs a set of measurements to mount namespaces.
The test is structured so that the host is measured first, followed by a
number of per-snap and per-user mount namespaces.

What the test doesn't do though, is to check that the test has not
affected the host in any way. This patch closes that loophole.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>