overlord: fix TestRemodelSwitchToDifferentKernel for bootvars
The bootvars tested in the TestRemodelSwitchToDifferentKernel are
not quite correctly mocked. When calling "state.MockRestarting()"
the code needs to clean the "snap_try_kernel" but also update
the "snap_kernel". This is what normally happens in
devicestate.DeviceManager.ensureBootOk() but the manager tests
won't run this so we need to simulate it.

This PR updates the bootvars correctly to make looking at the
test less confusing.