usersession/agent: don't manually send TERM/KILL when stopping services.
According to systemd.kill(5), systemd will automatically issue a kill
signal to processes remaining in the cgroup unless the unit sets
SendSIGKILL=no, which snapd does not do.

This apparently wasn't sufficient with ancient versions of systemd.  But
since this code will only run on systems with a working user instance of
systemd, we can ignore those versions.