snap-recovery: rename to "snap-bootstrap"
After a quick brainstorm with Gustavo and Samuele it was decided
to rename "snap-recovery" to something less specific. The binary
is primary about taking a disk and bootstraping partitions and
filesystems on it. So this PR renames it to snap-boostrap.

Note that this name may change again in the future but that is
fine because it's a internal name that is not directly user
visible (unless users poke into the initramfs).