packaging: tweak handling of usr.lib.snapd.snap-confine
This PR switches the code to use dpkg-maintscript-helper to deal
with usr.lib.snapd.snap-confine. Technically it's a dpkg conffile
(even though I think that is a mistake) so we should let the
dpkg tools handle it. This PR tweaks the call to try to remove it
regardless of version, it can cause quite a bit of harm to have
it, e.g. https://forum.snapcraft.io/t/13132

If it is acceptable to foundations I would like to unconditionally
rename it to ".dpkg-bak" but it's (technically) a violation of
the policy.