cmd/snap-confine: hide name=snapd cgroup from snaps
Ian noticed that the new cgroup breaks the docker snap. Docker
apparently tries to figure out where cgroups are and does something
unexpected when the new cgroup is in operation.

We had the same problem before, where docker would choke on the second
view of /sys via /var/lib/snapd/hostfs/sys. This patch fixes it with the
same approach, by hiding that view.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>