tests: add 14.04 canonical-livepatch test
This PR adds a new spread test that ensures that cannical-livepatch
works on 14.04 machines with re-exec and the snapd from the core
snap. We don't update the deb on 14.04 anymore but thanks to
re-exec 14.04 snapd is still fully up-to-date for people who
have the core snap installed. This PR ensures we don't break
this scenario.