tests: tweak wording in mount-ns test
Ian observed that the "before and after running the test" messages are a
little misleading since they are both used in the prepare section.  They
really meant to indicate before and after running application programs.
This is adjusted now.

In addition Maciej indicated that the extra line about persistent mount
namespace is not necessary as echo and can be demoted to a comment.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>