snapstate: fix undo of link-snap in kernel remodel
When we remodel to a new kernel we update the bootloader in the
"link-snap" task. If the remodel change needs to be undone this
change needs to be undone as well. Today we only undo bootloader
changes in the undo of the "unlink-current-snap". This was fine
in the past as any boot change would involve this task (e.g.
a refresh to a new kernel meant that the current kernel got the
"unlink-current-snap" task). However in a remodel add a new
kernel and never change the exiting one so no unlink-current-snap
is run (and hence no undo).

This PR fixes this by checking in the undoLinkSnap if we are
in a remodel and if so if the kernel is affected and if we
need to switch back.