packaging: tweak handling of usr.lib.snapd.snap-confine
* packaging: tweak handling of usr.lib.snapd.snap-confine

This PR switches the code to use dpkg-maintscript-helper to deal
with usr.lib.snapd.snap-confine. Technically it's a dpkg conffile
(even though I think that is a mistake) so we should let the
dpkg tools handle it. This PR tweaks the call to try to remove it
regardless of version, it can cause quite a bit of harm to have
it, e.g. https://forum.snapcraft.io/t/13132

If it is acceptable to foundations I would like to unconditionally
rename it to ".dpkg-bak" but it's (technically) a violation of
the policy.

* packaging: update snapd.maintscript on 14.04

We never changed /etc/apparmor.d/usr.lib.snapd.snap-confine to
snap-confine.real on 14.04. This means we cannot remove the
conffile there.

* packaging: use postinst snippet to get rid of usr.lib.snapd.snap-confine

This also reverts to the old rm_conffile handling, no changes
here.

Thanks to Steve for the suggestion.