overlord: add kernel rollback accross reboots manager test and fixes
This adds a test for kernel rollbacks accross reboots into the
managers test suite and fixes the behaviour. It seems like the
old code was not triggering a proper error and we did not notice
this in spread because we cannot replace an asserted kernel with
an unasserted so this test was incomplete.