interfaces/lxd-support: Fix on core18
The LXD interface effectively allows access to aa-exec.
This used to be in /usr/sbin/aa-exec on Ubuntu 16.04, on Ubuntu 18.04,
it's now in /usr/bin/aa-exec, making the interface unusable with core18.

This change allows aa-exec to be in any of:
 - /bin
 - /usr/bin
 - /sbin
 - /usr/sbin

Which should cover the location for core16, core18 and the upcoming core20.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>