tests/main/gadget-update-pc: cleanup per-revision directories
The test installs a repacked snaps, with forged revisions 1000-1002. However,
when the snapshot of / is restored after the test, files and only overwritten,
but the entries created by the test will not be removed. Make sure to clean up
the created directories manually when restoring, otherwise this may break the
tests that run later.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>