cmd/snap-confine: add sc_join_sub_group
This patch adds the logic for discovering the location in the unified
hierarchy and moving to a sub-directory of said hierarchy.

This allows all snaps, regardless of their type, "tag" their processes
in a way that is easy to detect. To detect presence of processes
belonging to a given snap one can lock the per-snap lock, iterate the
unified or v2 cgroup and look for all entries named snap.$SNAP_NAME.*.
Inside look at cgroup.procs and enumerate all processes.

This method works inside containers and across Ubuntu 14.04 all the way
to pure-cgroup v2 world Fedora 31. Given that it uses v2 in all cases,
we can leverage the cgroup.events file and notice vacancy of a cgroup
reliably, assuming snapd is running.

This patch adds the code and permissions but does not make it effective
yet.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>