many: use systemd scopes as identifiers
After discussing the issue of moving the process inside a cgroup
hierarchy with Zbyszek, the friendly, nearby systemd developer, we came
to the following conclusions:

- We cannot do this move without potentially confusing systemd
- We can ask systemd to do this move for us
- There is a pair of APIs that can be used to do this (see below)
- We need to use DBus for this
- There is no other way

The two APIs are StartTransientUnit and AttachProcessesToUnit. The first
one starts a new transient unit with at least one process inside. The
unit must have an unique name. The unit must be a scope or a service.
For what we are doing a scope makes more sense as it can be used by
services and doesn't break slice topology. The slice name is
"$SNAP_SECURITY_TAG.slice". If the scope already exists (there are
existing processes running there) then a different API is necessary,
namely AttachProcessesToUnit. This takes nearly the same arguments and
moves a process to the given unit, verifying consistency of ownership of
the caller, process and the unit. Attach has one more requirement, for
reasons unknown, that it can only attach processes to units that are
delegates. This means the scope that we start in StartTransientUnit
needs to be marked as a delegate.

Despite my earlier worry that using delegates would break service
management it doesn't seem to do so when using those APIs.

The implementation of DBus interaction is done with a huge hack, by
using system to spawn busctl to make the call. This will be adjusted
shortly.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>