boot,bootloader: make MakeBootable() uc20 aware
The current MakeBootable() code is pretty specific to UC16/UC18.
For UC20 we need to tweak it to install the recovery boot config
in boot.MakeBootable().

This commit adds a new recoveryGrub bootloader that is similar
to the real grub except that it operations with different dirs
and will never extract the kernels. We will need to modify the
"pc" gadget for UC20 to rename the "grub.conf" symlink to
"grub-recovery.conf" so that the bootloader can find the right
grub.cfg.

This will also allow to access the recovery bootloader config
via: `bootloader.Find("/run/mnt/ubuntu-seed")`. This will be
needed in e.g. PR#7762 where after the install from the
emphemeral mode we need to set the bootloader config of the
recovery grub to `snapd_recovery_mode=run`.