boot,bootloader: make MakeBootable() uc20 aware
The current MakeBootable() code is pretty specific to UC16/UC18.
For UC20 we need to tweak it to install the recovery boot config
in boot.MakeBootable().

This commit adds a new InstallBootCofig to the bootloader
interface. Most bootloader use the generic uc16 implementation
but for uc20 aware bootloader we will need to deal with the
recovery in a special way. We will also need to be able to
look at different base directories for grub (and potentially
the other bootloaders) as UC20 will have /boot/grub (system-boot)
and EFI/ubuntu (system-seed).

This will also allow to access the recovery bootloader config
via: `bootloader.Find("/run/mnt/ubuntu-seed")`. This will be
needed in e.g. PR#7762 where after the install from the
emphemeral mode we need to set the bootloader config of the
recovery grub to `snapd_recovery_mode=run`.