tests: run snap-set-core-config on all core devices
The snapctl-configure-core has a bit of a misleading name. It is
really about testing all the "snap set core " options. So this
commit renames/simplifies the test (we don't need a fake
bind mount anymore to test this). The new name is
snap-set-core-config.

The test runs now on all ubuntu-core-* devices. This means it
get properly tested on UC18 now too.

Note that this will only work once
https://github.com/snapcore/core18/pull/143 is merged and available
in the store.