cmd/snap-update-ns: implement new update-ns algorithm
When the robust-mount-namespace-updates feature is enabled mount
namespace changes are applied in a different way. Instead of a
differential algorithm we entirely undo the current mount profile and
then do the desired mount profile.

This has the benefit of eventually unmounting nested mimics correctly.

Tests are duplicated to check each variant, where applicable.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>