overlord/snapstate: add support for reverting the first install of snapd on core
The patch add support for failover handling and reverting of the first
installation of snapd on a core device. The snapstate backend UnlinkSnap() got a
new flag that indicates whether the unlink is part of first install
undo/cleanup, in which case it may need extra handling. So far the special
handling is only done for the snapd snap.

The intention is to use the code in a core to base remodel.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>