cmd/snap-failure: passthrough snapd logs, add informational logging (#7769)
When examining snapd recovery, it may be useful to be able to see the log output
generated by snapd. Currently, the output is captured and shown only when there
is an error.

Add some INFO level logs about steps taken by snap-failure, this is helpful when
investigating the logs manually.