interfaces/browser-support: allow writes to @{PROC}/@{pid}/clear_refs
Clearing the PG_Referenced and ACCESSED/YOUNG bits provides a method to
measure approximately how much memory a process is using via
/proc/self/smaps (man 5 proc). This access allows the snap to clear
references for pids from other snaps and the system, so it is limited to
snaps that specify: allow-sandbox: true. This is needed with newer
versions of chromium content api snaps.