devicestate: add device mapper warning to helper
The helper that determines the disk from a partition only works with
device nodes corresponding to real disk partitions and will fail if
we use a device-mapped block device (also because if using the device
mapper it can map to different physical disks). Add a warning to
explicit this limitation.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>