overlord/snapstate: pick up system defaults when seeding the snapd snap
On a system with bases, the snapd snap is considered a 'system' snap, thus we
consider it as source of system configuration defaults. Capture this in the
code, make sure that system defaults are only applied when in the seeding mode.
Distinguish cases when both snapd and core snaps are installer, or just one of
those.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>