update system-backup tests to not check for sanitize errors related to os
PR 6950 removed the need for checking for sanitize errors related to os
and gadget snaps, but PR 6436 which added the system-backup interface
was not updated (and the tests hadn't been run since PR 6950 was
committed). This brings this interface's checks in line with all other
interfaces that have a similar setup (implicit core and classic with
same base declaration).