overlord/snapstate: do not trigger defaults when installing the core snap
Do not trigger the defaults when installing the core snap. This is only done
during seeing, via an explicitly added configure 'core' task.

Otherwise, this leads to a scenarion when on a core device with bases, the
defaults are applied once during seeding, and then again when the core snap is
being installed. This may cause the system config flags to switch back to their
default values.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>