boot: add boot.Modeenv.Kernel support
For the grub bootloader on the ubuntu-boot partition we want to
keep the kernel commandline as static as possible to avoid TPM
reseal operations for each kernel update.

This means that we will have a kernel with a fixed name on the
ubuntu-boot partition. But the initramfs needs information what
kernel snap we expect so that it can mount the kernel snap
and make the kernel modules available early in initramfs.

This commit adds "kernel" to the modeenv so that snapd can
set what kernel snap is expected via the modeenv mechanism
that is already used to set what base to use. The snap-bootstrap
initramfs-mounts code will have to do a cross-check to ensure
that the booted kernel matches the expected kernel and deal
with kernel reverts from the bootloader. This will happen in
subsequent commits.