testutil: add a locked variant of mocked command
The script generated by MockCommand() writes the call list into the log. Thus,
it cannot be called in parallel without risking the call log being corrupted by
interleaving writes.

Add a locked variant of mocked command that wraps the script execution with a
flock call that takes an exclusive lock.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>