snap/squashfs: verify files/dirs can be accessed by mksquashfs when building a snap
The mksquashfs tool does not error out when locations are not accessible by the
user who invoked it. Workaround that by adding an early pass that verifies the
contents of snap directory are accessible (as in R_OK) to the current user, and
produce an error otherwise.

For instance, attempting to pack the core18 snap as a user produces this error
message:

```
error: cannot pack "/snap/core18/current": cannot access the following locations in the snap source directory:
- etc/gshadow (owner 0:42 mode 0640)
- etc/shadow (owner 0:42 mode 0640)
- etc/ssl/private (owner 0:0 mode 0700)
- etc/sudoers (owner 0:0 mode 0440)
- etc/sudoers.d/README (owner 0:0 mode 0440)
- root (owner 0:0 mode 0700)
- var/cache/ldconfig (owner 0:0 mode 0700)
- var/lib/private (owner 0:0 mode 0700)
- var/lib/snapd/void (owner 0:0 mode 0111)
```

Fixes: https://bugs.launchpad.net/snapd/+bug/1854045

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>