devicestate: call boot.MakeRunnable() in devicestate
When installing a UC20 system from the recovery partition snapd
needs to make the run system runnable. This is done via a new
boot.MakeRunnable() helper that will setup the required bootloader
and modeenv changes. This commit adds an empty MakeRunnable()
helper and the high level wiring so that the helper is called
in the right way (plus tests for this).