interfaces/builtin: add uio interface
The UIO interface allows for userspace to implement device drivers for
specific hardware. It is documented in the kernel, for example here:
https://www.kernel.org/doc/html/v4.12/driver-api/uio-howto.html

As implemented the slot can be either statically defined by the gadget
snap, where it must contain a path attribute that points to
/dev/uio[0-9]+ device name, or can be dynamically detected by snapd
hotplug system, if enabled.

I've successfully tested the interface on a customer device.

Fixes: https://bugs.launchpad.net/snapd/+bug/1641752
Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>