boot,bootlaoder: setup the snap recovery system bootenv
When making a UC20 system bootable from `snap prepare-image` a
bootenv that is specific to the recovery system needs to get
written so that the bootloader can find the right kernel for
the given recovery system.

This is now done via a new bootloader.RecoveryAwareBootloader
interface that will write recovery system specific details
to the bootloader configuration.