interfaces: add raw-volume interface for access to partitions (#7768)
While the block-devices interface provides access to raw disks, some
applications need access to disk partitions at the block level. This
interface provides this access. Since read access gives the snap access
to all data on the device and write access allows changing the contents
of the disk partition, this interface is considered a "super-privileged"
interface since it grants device ownership to the snap.

This is needed for enabling certain UC20 devices and needs to be in 2.43.