wrappers: improve robustness of installing snapd snap on core
When the snapd snap is installed for the first time, we will write the unit
files into /etc/systemd/{system,user}. There is no need to reenable the services
that are already enabled. The symlink in /etc/systemd/system/*.wants/* does not
need to point to the /etc file for the right unit to be used. Instead systemd,
handles the units logically, with /etc being of higher priority than other /usr,
or /lib locations, the right unit file will be used.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>