data: adjust systemd ConditionKernelCommandLine for UC20
The gadget snap for uc20 got updated recently and no longer
includes "snap_core=" in the kernel commandline. This means
that some of the services that are supposed to run on all
core devices do not run on UC20 today.

This PR updates the check so that snapd_recovery_mode is also
recognized as a valid ubuntu core commandline.