cmd/libsnap: add sc_mk{,sub}dir for creating directories
The new pair of function assists in creating directories and
sub-directories with correct ownership and permissions, avoiding race
conditions that would allow non-root users to open it during the
intermediate window between mkdir and chown.

In addition, the utils-test.c file now only includes utils.c, not the
combination of the .c and .h files. This allows the .c file to define
the full set of preprocessor macros that should govern compilation, such
as defining _GNU_SOURCE.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>