cmd/snap-confine: run with effective gid of calling user
This change addresses the main point raised by the security review of my
earlier version of this patch, intending to correct and somehow simplify
the act of creating files and directories so that we don't create them
with the group ownership of the calling user.

The change reverses the direction, by running with effective GID of the
user calling into snap-confine. Previously I wanted to use effective GID
of zero, so that all directories would be root owned automatically. This
is now compensated by helpers around mkdir. Patches prior in this series
change all usage of mkdir to sc_mkdir or sc_mksubdir, so that this issue
is still resolved.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>