snap: default to "--direct" in `snap known` for 2.43
This commit changes the default in `snap known --remote` to a direct
store query again. It remove the ability to query remote assertions
via snapd for now. This is to be consistent with the `snap download`
command which also goes directly to the store. For 2.44 we will
switch snap download to download via the snapd daemon.