snap: remove "host" output from `snap version`
The newly added output of `host $cpu $virtualization` broke existing
software that relied on the fact that `snap version` has a single
whitespace delimiter.

The "host" output was added to help people give us good
information for bugreport. I.e. the "snap version" output should
be part of each bugreport. But conceptually "host" is not a good
fit for the version output so this commit remove it again.

In the future we will provide a "snap debug" command that will
output the relevant version information plus the architecture
and the virtualization environment.