boot: add HasModeenv to Device
This will be used by bootStateFor in determining if a given device is a UC20
device and thus needs a different bootState implementation.

Adjust some tests to ensure that we have a modeenv when expected and don't when
expected.

This also requires refactoring the MockDevice in boottest to be slightly more
complicated to have devices that are in run mode, but are not uc20.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>