bootloader: add ExtractedRunKernelImageBootloader interface, implement in grub
This is used for UC20, when we need to manipulate kernel assets through some
bootloader state not in the variables. Specifically, we need methods to
manipulate a kernel.efi symlink from the root of the ubuntu-boot partition to
the extracted kernel.efi asset for the current kernel snap, as well as another
symlink for kernel updates, the try-kernel.efi symlink. We need to be able to
add these symlinks, remove them, and inspect them.

Also add numerous testing functions to the mock bootloader to make testing that
code calls correct things when manipulating and inspecting these things in the
bootloader.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>