boot: implement bootState20
This implements the bootStateUpdate and bootState interfaces from boot the way
we want for UC20. Currently only kernel snaps are handled, with the minimal
amount of work to enable GetCurrentBoot() to work for base snaps as well (though
base snaps being updated are not supported here yet).

We use the ExtractedRunKernelImageBootloader interface to manipulate and inspect
the various kernel assets like kernel.efi and try-kernel.efi symlinks.

The commit() method should be complete enough to implement robustness where
reboots at any point in execution should not block a successful kernel reboot,
though it would be good if we had spread tests to confirm this.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>