grub: fix EnableKernel() and EnableTryKernel() symlink locations
When we are creating the symlinks, we should always create relative symlinks 
because for example when ubuntu-boot grub is mounted at /run/mnt/ubuntu-boot, 
then absolute kernel.efi symlinks to /EFI/ubuntu/<snap>/kernel.efi will not 
resolve, but a relative one will. 

Also refactor the checks a bit to use Lstat where appropriate where we just want
to test if there is a symlink there and check if the symlink is not dangling 
later.

Also fix bug that always used "kernel.efi" instead of the name argument.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>