render: add the render package and basic widgets
The render package allows compositing textual widgets and display them
in a terminal effortlessly. Unlike hand-crafted wrap-and-print-spaces
code it is easier to handle complex structures correctly.

Rendering is based on stripes. The widget tree is de-composed into
horizontal stripes, or rectangles of height of one, and rendered into a
single-line temporary buffer using the painter's algorithm. There's some
provisions for supporting double-width characters, allowing to mix i18n
text with arbitrary English text correctly.

The widgets sub-package contains the following widgets:
 Pre - pre-formatted text, text label that is similar to Printf
 Padding - container that applies padding to another widget
 HBox and VBox - containers that pack widgets horizontally and
 vertically.

The heuristics sub-package contains code that guesses what a terminal
would do with a given string, mainly to determine the width and height
of the encompassing rectangle.

For simplicity, all of this is wrapped into simple functions that craft
structural presentation: Sequences, Headers, Text, Lists and Maps.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>